requirements.txt
plexapi
coloredlogs
python-dotenv